helpers: introduce render_w_mentions
Eventually this function should support and auto detect multiple formats and is
thus not named for a specific format. But for now it is plain text only.

This kind of markup can quite easily and safely support additional magic
markup. It is much harder to do that on top of a richer markup format; it
must essentially be done in a single pass, with both all the various regexps
and the rst formatting done in a single pass.