hgicon.png: use repotag class to make an icon
The inline 'style' attribute was modified to remove the color but nothing else,
as was discussed with other Kallithea devs, since it makes more sense to leave
the spacing until another pass of css refactoring.

A quick search revealed that some old css code was never used, so it was
removed.