diff view: do not hide overflow on very long lines that cannot be wrapped
When a change contains very long lines that cannot be wrapped because they
contain very long 'words' without spaces, the overflow is currently not
visible at all. This occurrence of such very long non-wrappable words is
pretty uncommon, but can for example happen in comma-separated value files.

Not touching the existing behavior of trying to wrap long lines, this patch
adds horizontal scrollbars for these exceptional cases of unavoidable
overflow.
On normal lines and long lines that can be wrapped, the diff block remains
the same as before.