diff: get collapse target via .attr instead of .prop
Commit 3f017db297c4 was not fully tested and broke collapse/expand of diffs on
changesets. $button is not a link with a target and the target can thus not be
retrieved with .prop('target'); $button is just a span that happens to have a
custom attribute with the name 'target'.

We thus revert back to the old way of retrieving it with .attr('target'). (It
would perhaps be even better to use data attributes and name it data-target and
use .data('target') ...)