auth: Fix bug where usernames are not consistently capitalized when using crowd login
If you try to log in to Kallithea via the Crowd auth module then the
capitalization of your username in Kallithea changes on every login based on
how you capitalized it in the login form.

E.g. Log in with "TestDude", username is entered as "TestDude" then log in
again, but this time as "tesTduDe", and your username gets changed to
"tesTduDe". etc.

Fix for this is to use the 'name' field returned from Crowd when saving the
username. This way the username is always capitalized identically to the record
in Crowd.